# backend/openapi.yaml
openapi: 3.0.3
info:
  title: SAuth API
  version: 1.0.0
  description: API specification for SAuth authentication flow
servers:
  - url: https://sauth.example.com/api
tags:
  - name: SAuth Admin
    description: APIs for admin backend
  - name: SAuth Web
    description: APIs for SAuth Web (QR Page)
  - name: SAuth Mobile
    description: APIs for SAuth Mobile app
  - name: Client
    description: APIs for client backend
  - name: Client Webhook
    description: APIs for client backend Webhook
paths:
  # --- Client APIs ---
  /client:
    post:
      tags: [SAuth Admin]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                webhook_url:
                  type: string
                enable_datadome:
                  type: boolean
                enable_recaptcha:
                  type: boolean
              required:
                - name
                - webhook_url
      responses:
        "200":
          description: Registered client
          content:
            application/json:
              schema:
                type: object
                properties:
                  client_id:
                    type: string
                  client_secret:
                    type: string
    put:
      tags: [SAuth Admin]
      summary: Update existing client
      parameters:
        - name: client_id
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                webhook_url:
                  type: string
                enable_datadome:
                  type: boolean
                enable_recaptcha:
                  type: boolean
              required:
                - name
                - webhook_url
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

    get:
      tags: [SAuth Admin]
      summary: Get all clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    client_id:
                      type: string
                    client_secret:
                      type: string
                    api_key:
                      type: string
                    name:
                      type: string
                    webhook_url:
                      type: string
                    enable_datadome:
                      type: boolean
                    enable_recaptcha:
                      type: boolean
                    created_at:
                      type: string
  /client/{clientId}:
    delete:
      tags: [SAuth Admin]
      summary: Delete a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
          description: The client ID to delete
      responses:
        "200":
          description: Client deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  # --- SAuth Web APIs ---
  /qrcode/{request_id}:
    get:
      tags: [SAuth Web]
      summary: SAuthWeb requests QR code from SAuthBackend
      parameters:
        - name: request_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: QR code data
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                  expired_time:
                    type: string
                  nonce:
                    type: string
                  allow_retry:
                    type: boolean

  # --- SAuth Mobile APIs ---
  /mobile/register-device:
    post:
      tags: [SAuth Mobile]
      summary: SAuth Mobile registers a new device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                public_key:
                  type: string
                signature:
                  type: string
                device_info:
                  type: object
              required:
                - public_key
                - signature
      responses:
        "200":
          description: Device registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  device_id:
                    type: string
  /mobile/verify-qrcode:
    post:
      tags: [SAuth Mobile]
      summary: SAuth Mobile verifies QR code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                nonce:
                  type: string
                signature:
                  type: string
                device_id:
                  type: string
              required:
                - session_id
                - nonce
                - signature
                - device_id
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
  # --- Client APIs ---
  /verify-request:
    post:
      tags: [Client]
      summary: Send verify request
      parameters:
        - name: client_id
          in: header
          required: true
          schema:
            type: string
        - name: client_secret
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                allow_refresh:
                  type: boolean
                expired_duration:
                  type: number
                max_try:
                  type: number
                metadata:
                  type: object
              required:
                - allow_refresh
                - expired_duration
                - max_try
      responses:
        "200":
          description: Verification request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                  request_id:
                    type: string
  /webhook:
    post:
      tags: [Client Webhook]
      summary: Webhook for client to receive verification result
      parameters:
        - name: api_key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                request_id:
                  type: string
                status:
                  type: string
                device_id:
                  type: string
                metadata:
                  type: object
              required:
                - request_id
                - status
                - device_id
      responses:
        "200":
          description: Webhook received
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string